+++ util/opentelemetry-util-genai/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- Enrich ToolCall type, breaking change: usage of ToolCall class renamed to ToolCallRequest 
+- Enrich ToolCall type, breaking change: usage of ToolCall class renamed to ToolCallRequest
   ([#4218](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4218))
 - Add EmbeddingInvocation span lifecycle support
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4219](#4219))
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4320](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4320))
 - Add workflow invocation type to genAI utils
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4310](#4310))
+- Allow passing external Span to the manual LLMInvocation
+  ([#4281](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4281))
 
 ## Version 0.3b0 (2026-02-20)